+++ emote/callbacks/__init__.py
@@ -4,8 +4,9 @@
 
 from emote.callbacks.checkpointing import Checkpointer, CheckpointLoader
 from emote.callbacks.generic import BackPropStepsTerminator
-from emote.callbacks.logging import LoggingMixin, TensorboardLogger
+from emote.callbacks.logging import TensorboardLogger
 from emote.callbacks.loss import LossCallback
+from emote.mixins.logging import LoggingMixin
 
 
 __all__ = [